Виконує завдання
