--- controllers/auth.js
@@ -1,94 +1,0 @@
-const { Conflict, Unauthorized } = require("http-errors");
-
-const jwt = require("jsonwebtoken");
-const bcrypt = require("bcryptjs");
-require("dotenv").config();
-
-const { User } = require("../models/user");
-const { ctrlWrapper } = require("../helpers");
-
-const { SECRET_KEY } = process.env;
-
-const register = async (req, res) => {
-  const { email, password, subscription } = req.body;
-  const user = await User.findOne({ email });
-
-  if (user) {
-    throw new Conflict(`Email ${email} in use`);
-  }
-  const hashPassword = await bcrypt.hash(password, 10);
-  const newUser = await User.create({
-    email,
-    password: hashPassword,
-    subscription,
-  });
-
-  res.status(201).json({
-    user: {
-      email: newUser.email,
-      subscription: newUser.subscription,
-    },
-  });
-};
-
-const login = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-
-  if (!user) {
-    throw new Unauthorized("Email or password is wrong");
-  }
-  const passwordCompare = await bcrypt.compare(password, user.password);
-  if (!passwordCompare) {
-    throw new Unauthorized("Email or password is wrong");
-  }
-
-  const payload = {
-    id: user._id,
-  };
-  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
-  await User.findByIdAndUpdate(user._id, { token });
-  res.status(200).json({
-    user: {
-      email,
-      subscription: user.subscription,
-    },
-    token,
-  });
-};
-
-const getCurrent = async (req, res) => {
-  const { email, subscription } = req.user;
-  res.json({
-    email,
-    subscription,
-  });
-};
-
-const logout = async (req, res) => {
-  const { _id } = req.user;
-  await User.findByIdAndUpdate(_id, { token: "" });
-  res.status(204).json();
-};
-
-const updateSubscription = async (req, res) => {
-  console.log(req.user);
-  const { _id } = req.user;
-  const { subscription } = req.body;
-  const changedSubscription = await User.findByIdAndUpdate(
-    _id,
-    { subscription },
-    {
-      new: true,
-    }
-  );
-  res.status(200).json(changedSubscription);
-};
-
-module.exports = {
-  register: ctrlWrapper(register),
-  login: ctrlWrapper(login),
-  getCurrent: ctrlWrapper(getCurrent),
-  logout: ctrlWrapper(logout),
-  updateSubscription: ctrlWrapper(updateSubscription),
-};
