+++ controllers/auth/auth.js
@@ -1,0 +1,16 @@
+const { ctrlWrapper } = require("../../helpers");
+const register = require("./register");
+const login = require("./login");
+const getCurrent = require("./getCurrent");
+const logout = require("./logout");
+const updateSubscription = require("./updateSubscription");
+const updateAvatar = require("./updateAvatar");
+
+module.exports = {
+  register: ctrlWrapper(register),
+  login: ctrlWrapper(login),
+  getCurrent: ctrlWrapper(getCurrent),
+  logout: ctrlWrapper(logout),
+  updateSubscription: ctrlWrapper(updateSubscription),
+  updateAvatar: ctrlWrapper(updateAvatar),
+};
