+++ controllers/auth/login.js
@@ -1,0 +1,36 @@
+const { Unauthorized } = require("http-errors");
+const jwt = require("jsonwebtoken");
+const bcrypt = require("bcryptjs");
+require("dotenv").config();
+
+const { User } = require("../../models/user");
+
+const { SECRET_KEY } = process.env;
+
+const login = async (req, res) => {
+  const { email, password } = req.body;
+  const user = await User.findOne({ email });
+
+  if (!user) {
+    throw new Unauthorized("Email or password is wrong");
+  }
+  const passwordCompare = await bcrypt.compare(password, user.password);
+  if (!passwordCompare) {
+    throw new Unauthorized("Email or password is wrong");
+  }
+
+  const payload = {
+    id: user._id,
+  };
+  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
+  await User.findByIdAndUpdate(user._id, { token });
+  res.status(200).json({
+    user: {
+      email,
+      subscription: user.subscription,
+    },
+    token,
+  });
+};
+
+module.exports = login;
