+++ controllers/auth/register.js
@@ -1,0 +1,31 @@
+const { Conflict } = require("http-errors");
+const bcrypt = require("bcryptjs");
+require("dotenv").config();
+const gravatar = require("gravatar");
+const { User } = require("../../models/user");
+
+const register = async (req, res) => {
+  const { email, password, subscription } = req.body;
+  const user = await User.findOne({ email });
+
+  if (user) {
+    throw new Conflict(`Email ${email} in use`);
+  }
+  const hashPassword = await bcrypt.hash(password, 10);
+  const avatarURL = gravatar.url(email);
+  const newUser = await User.create({
+    email,
+    password: hashPassword,
+    subscription,
+    avatarURL,
+  });
+
+  res.status(201).json({
+    user: {
+      email: newUser.email,
+      subscription: newUser.subscription,
+    },
+  });
+};
+
+module.exports = register;
