+++ controllers/contacts/contacts.js
@@ -1,0 +1,17 @@
+const { ctrlWrapper } = require("../../helpers");
+
+const listContacts = require("./listContacts.js");
+const getContactById = require("./getContactById");
+const addContact = require("./addContact");
+const deleteContact = require("./deleteContact");
+const updateContact = require("./updateContact");
+const updateFavourite = require("./updateFavourite");
+
+module.exports = {
+  listContacts: ctrlWrapper(listContacts),
+  getContactById: ctrlWrapper(getContactById),
+  addContact: ctrlWrapper(addContact),
+  deleteContact: ctrlWrapper(deleteContact),
+  updateContact: ctrlWrapper(updateContact),
+  updateFavourite: ctrlWrapper(updateFavourite),
+};
