+++ middlewares/index.js
@@ -1,9 +1,11 @@
 const validateBody = require("./validateBody");
 const isValidId = require("./isValidId");
 const authentificate = require("./authentificate");
+const upload = require("./upload");
 
 module.exports = {
   validateBody,
   isValidId,
   authentificate,
+  upload,
 };
