+++ package.json
@@ -14,12 +14,15 @@
     "cross-env": "7.0.3",
     "dotenv": "^16.1.4",
     "express": "4.17.1",
+    "gravatar": "^1.8.2",
     "http-errors": "^2.0.0",
+    "jimp": "^0.22.8",
     "joi": "^17.9.2",
     "jsonwebtoken": "^9.0.0",
     "mongodb": "^5.6.0",
     "mongoose": "^7.2.3",
     "morgan": "1.10.0",
+    "multer": "^1.4.5-lts.1",
     "nanoid": "^3.3.4"
   },
   "devDependencies": {
